+++ source/os/console.ts
@@ -93,7 +93,8 @@ module TSOS {
                     this.setBuffer(newBuff)
                 } else if (chr === "ctrl-c") {
 
-                    _Scheduler.killProcess()
+                    if (_CPU.isExecuting)
+                        _Scheduler.killProcess()
 
                     this.historyIndex = 0
                     this.buffer = ""
